menusystem/requirements/settings.txt
Django==4.0.4
django-cors-headers==3.11.0
mssql-django==1.1.3
pytz==2022.1
django-semantic-admin==0.3.2
djangorestframework==3.13.1
django-admin-sortable2==2.1
dj-database-url==0.5.0